django_project/requirements.txt
Django==2.2.4
django-bootstrap4==0.0.8
django-tables2-bootstrap4==2.0.0b5
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
